Outsource code of finding candidates to new file
