+++ src/lib.rs
@@ -37,6 +37,7 @@ pub mod cli;
 pub(crate) mod conversion;
 pub(crate) mod errors;
 pub(crate) mod estimation;
+pub (crate) mod candidates;
 pub mod filtration;
 pub(crate) mod grammar;
 pub(crate) mod reference;
